Fix #305
git-svn-id: https://build.archivesportaleurope.net/svn/apex-dpt/branches/DPT-project-1.2.0@161 87c78061-3d99-4f1f-8d5c-9a005f5ba2cf

+++ DPTstandalone/src/main/java/eu/apenet/dpt/standalone/gui/XsltAdderActionListener.java
@@ -59,10 +59,14 @@ public class XsltAdderActionListener implements ActionListener {
 
     private boolean saveXslt(File file){
         try{
+            if(new File(Utilities.CONFIG_DIR + file.getName()).exists())
+                throw new Exception("XSL file already exists");
             FileUtils.copyFile(file, new File(Utilities.CONFIG_DIR + file.getName()));
             return true;
         } catch (IOException e){
             LOG.error("Moving file " + file.getAbsolutePath() + " failed", e);
+            return false;
+        } catch (Exception e) {
             return false;
         }
     }
